// generated/src/app/api-models/general-ledger-assign.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class GeneralLedgerAssign {
  seqGlAssign: number;
  glassKey1: string;
  glassKey2: string;
  glassKey3: string;
  glassKey4: string;
  glassKey5: string;
  glassKey6: string;
  companyCode: string;
  glRefCode: string;
  description: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}